Add emacs lisp test runner script
I've been toying around with test driven emacs lisp. Using ert inside
of emacs is nice, but over time it becomes difficult to reason about
the state of the emacs process and how it MAY be affecting the
pass/fail state of tests. This script allows me to start a fresh,
isolated instance of the emacs process in which I can run the tests to
avoid contaminated results.