Add ablity to port around gnome-terminal profiles
Previously, this configuration only cared about the gruvbox color
scheme. And for the most part that kinda stays the same. However, this
change is intended to preserve ALL of the gnome-terminal profiles I
use. The next step would be to automatically install them when
installing the .files on GNU/Linux hosts.